Form need's item

+++ app/src/main/java/com/example/gabriela/agenda/model/Classmate.java
@@ -51,7 +51,7 @@ public class Classmate implements Serializable{
         this.site = site;
     }
 
-    public Double getStars() {
+    public int getStars() {
         return stars;
     }
